Refactor AI insights types and remove unused EmailSignInForm
- Moved all AI insights related TypeScript interfaces from a separate types file into the main ai-insights-screen.tsx component for better encapsulation and maintainability.
- Deleted the now-redundant .
- Removed the unused  component to clean up the codebase.
